Split out manifest and lock into their own files
